application/requirements.txt
Flask==2.3.2
MarkupSafe==2.1.2
numpy==1.23.5
opencv_contrib_python_headless==4.7.0.72
opencv_python==4.7.0.72
pandas==2.0.1
Pillow==9.5.0
pydicom==2.3.1
tensorflow==2.12.0
waitress==2.1.2
